refactor: extract generate_impl_into_dart_related (general)
