+++ frb_codegen/src/library/codegen/generator/wire/rust/spec_generator/codec/dco/encoder/misc.rs
@@ -1,4 +1,20 @@
+use crate::codegen::generator::codec::structs::CodecMode;
+use crate::codegen::generator::misc::comments::generate_codec_comments;
 use crate::codegen::ir::namespace::NamespacedName;
+
+pub(crate) fn generate_impl_into_dart(name: &str, body: &str) -> String {
+    let codec_comments = generate_codec_comments(CodecMode::Dco);
+    format!(
+        "{codec_comments}
+            impl flutter_rust_bridge::IntoDart for {name} {{
+                fn into_dart(self) -> flutter_rust_bridge::for_generated::DartAbi {{
+                    {body}
+                }}
+            }}
+            impl flutter_rust_bridge::for_generated::IntoDartExceptPrimitive for {name} {{}}
+"
+    )
+}
 
 pub(crate) fn generate_impl_into_into_dart(
     name: &NamespacedName,
@@ -18,6 +34,7 @@ pub(crate) fn generate_impl_into_into_dart(
             fn into_into_dart(self) -> {wrapper_name} {{
                 {body}
             }}
-        }}"
+        }}
+"
     )
 }
